Create p45 device module holding just classes
